chart view dart file cleaned
