+++ lib/Components/Tab_Views/chart_view/barchart_model.dart
@@ -1,0 +1,7 @@
+class Pollution {
+  String place;
+  int year;
+  int quantity;
+
+  Pollution(this.year, this.place, this.quantity);
+}
