+++ lib/Components/Tab_Views/chart_view/pieChart_model.dart
@@ -1,0 +1,9 @@
+import 'dart:ui';
+
+class Task {
+  String task;
+  double taskvalue;
+  Color colorval;
+
+  Task(this.task, this.taskvalue, this.colorval);
+}
